move the ui library directly to the extension
