+++ packages/extension/package.json
@@ -25,18 +25,26 @@
     "zip-a-folder": "^1.1.5"
   },
   "dependencies": {
+    "@otonashixav/solid-flip": "^0.10.5",
     "@solid-devtools/shared": "workspace:^0.7.3",
-    "@solid-devtools/ui": "workspace:0.10.2",
     "@solid-primitives/context": "^0.1.1",
+    "@solid-primitives/cursor": "^0.0.100",
     "@solid-primitives/destructure": "^0.1.2",
+    "@solid-primitives/event-listener": "^2.2.2",
     "@solid-primitives/immutable": "^0.1.2",
+    "@solid-primitives/keyboard": "^1.0.2",
     "@solid-primitives/keyed": "^1.1.2",
     "@solid-primitives/props": "^2.2.2",
+    "@solid-primitives/resize-observer": "^2.0.4",
     "@solid-primitives/rootless": "^1.1.3",
     "@solid-primitives/set": "^0.2.2",
+    "@solid-primitives/timer": "^1.3.2",
     "@solid-primitives/utils": "^3.0.2",
     "@vanilla-extract/css": "^1.9.0",
+    "@vanilla-extract/dynamic": "^2.0.2",
     "@vanilla-extract/private": "^1.0.3",
+    "clsx": "^1.2.1",
+    "csstype": "^3.1.0",
     "solid-js": "^1.5.4",
     "solid-transition-group": "^0.0.11",
     "type-fest": "^2.19.0"
