+++ packages/extension/src/index.tsx
@@ -4,8 +4,6 @@ import "./state/graph"
 import "./state/details"
 import "./bridge"
 
-import "@solid-devtools/ui/css"
-
 import App from "./App"
 
 render(() => <App />, document.getElementById("root")!)
