+++ packages/extension/src/ui/components/splitter/Splitter.tsx
