+++ packages/extension/src/ui/index.ts
@@ -8,4 +8,6 @@ export { Skeleton } from "./components/loading/Skeleton"
 export { OwnerPath } from "./components/owner/Path"
 export * from "./components/button/Button"
 
+export * as Icon from "./icons"
+
 export * from "./ctx/structure"
