--- packages/ui/package.json
@@ -1,101 +1,0 @@
-{
-  "name": "@solid-devtools/ui",
-  "version": "0.10.2",
-  "description": "UI Library for visualization of the reactive graph. Powering Solid Devtools.",
-  "license": "MIT",
-  "author": "Damian Tarnawski <gthetarnav@gmail.com>",
-  "contributors": [],
-  "homepage": "https://github.com/thetarnav/solid-devtools/tree/main/packages/ui#readme",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/thetarnav/solid-devtools.git"
-  },
-  "bugs": {
-    "url": "https://github.com/thetarnav/solid-devtools/issues"
-  },
-  "keywords": [
-    "solid",
-    "devtools"
-  ],
-  "private": false,
-  "sideEffects": true,
-  "type": "module",
-  "main": "dist/index.cjs",
-  "module": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "exports": {
-    ".": {
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs"
-    },
-    "./css": {
-      "import": "./dist/index.css",
-      "require": "./dist/index.css"
-    },
-    "./theme": {
-      "import": "./dist/theme/index.js",
-      "require": "./dist/theme/index.cjs"
-    },
-    "./animation": {
-      "import": "./dist/animation/index.js",
-      "require": "./dist/animation/index.cjs"
-    },
-    "./icons": {
-      "import": "./dist/icons/index.js",
-      "require": "./dist/icons/index.cjs"
-    }
-  },
-  "typesVersions": {
-    "*": {
-      "theme": [
-        "./dist/theme/index.d.ts"
-      ],
-      "animation": [
-        "./dist/animation/index.d.ts"
-      ],
-      "icons": [
-        "./dist/icons/index.d.ts"
-      ]
-    }
-  },
-  "scripts": {
-    "dev": "tsup --watch",
-    "build": "tsup",
-    "test": "echo \"NOOP TEST\"",
-    "typecheck": "tsc --noEmit"
-  },
-  "devDependencies": {
-    "@solid-devtools/transform": "workspace:^0.7.2",
-    "@vanilla-extract/esbuild-plugin": "^2.1.0",
-    "esbuild": "^0.15.7",
-    "esbuild-plugin-solid": "^0.4.2",
-    "solid-js": "^1.5.4",
-    "tsup": "^6.2.3",
-    "typescript": "^4.8.2"
-  },
-  "dependencies": {
-    "@otonashixav/solid-flip": "^0.10.5",
-    "@solid-aria/button": "^0.1.3",
-    "@solid-aria/interactions": "^0.1.4",
-    "@solid-devtools/shared": "workspace:^0.7.3",
-    "@solid-primitives/bounds": "^0.0.102",
-    "@solid-primitives/cursor": "^0.0.100",
-    "@solid-primitives/event-listener": "^2.2.2",
-    "@solid-primitives/keyboard": "^1.0.2",
-    "@solid-primitives/keyed": "^1.1.2",
-    "@solid-primitives/props": "^2.2.2",
-    "@solid-primitives/resize-observer": "^2.0.4",
-    "@solid-primitives/timer": "^1.3.2",
-    "@solid-primitives/utils": "^3.0.2",
-    "@vanilla-extract/css": "^1.9.0",
-    "@vanilla-extract/dynamic": "^2.0.2",
-    "@vanilla-extract/private": "^1.0.3",
-    "clsx": "^1.2.1",
-    "csstype": "^3.1.0",
-    "solid-transition-group": "^0.0.11"
-  },
-  "peerDependencies": {
-    "solid-js": "^1.5.0"
-  },
-  "packageManager": "pnpm@7.11.0"
-}
